refactor management of ad free cookie and reasons
move getting, setting, management of reasons to manage-ad-free-cookie.ts
