+++ static/src/javascripts/projects/common/modules/commercial/user-features.ts
@@ -1,4 +1,12 @@
 import { getCookie, isObject, removeCookie, setCookie } from '@guardian/libs';
+import {
+	AdFreeCookieReasons,
+	adFreeDataIsOld,
+	adFreeDataIsPresent,
+	getAdFreeCookie,
+	maybeUnsetAdFreeCookie,
+	setAdFreeCookie,
+} from 'lib/manage-ad-free-cookie';
 import { noop } from 'lib/noop';
 import config from '../../../../lib/config';
 import { fetchJson } from '../../../../lib/fetch-json';
@@ -7,11 +15,11 @@ import { getLocalDate } from '../../../../types/dates';
 import type { LocalDate } from '../../../../types/dates';
 import type { UserFeaturesResponse } from '../../../../types/membership';
 import { isUserLoggedIn } from '../identity/api';
+import { cookieIsExpiredOrMissing, timeInDaysFromNow } from './utils';
 
 // Persistence keys
 const USER_FEATURES_EXPIRY_COOKIE = 'gu_user_features_expiry';
 const PAYING_MEMBER_COOKIE = 'gu_paying_member';
-const AD_FREE_USER_COOKIE = 'GU_AF1';
 const ACTION_REQUIRED_FOR_COOKIE = 'gu_action_required_for';
 const DIGITAL_SUBSCRIBER_COOKIE = 'gu_digital_subscriber';
 const HIDE_SUPPORT_MESSAGING_COOKIE = 'gu_hide_support_messaging';
@@ -39,24 +47,18 @@ const CONTRIBUTIONS_REMINDER_SIGNED_UP = {
 	daysToLive: 90,
 };
 
-const AD_FREE_COOKIE_REASON_LS = 'gu.ad_free_cookie_reason';
-enum AdFreeCookieReasons {
-	AdFreeCookieReasonUserOptOut = 'user_opt_out',
-	AdFreeCookieReasonSubscriber = 'subscriber',
-}
-
 // TODO: isn’t this duplicated from commercial features?
 // https://github.com/guardian/frontend/blob/2a222cfb77748aa1140e19adca10bfc688fe6cad/static/src/javascripts/projects/common/modules/commercial/commercial-features.ts
 const forcedAdFreeMode = !!/[#&]noadsaf(&.*)?$/.exec(window.location.hash);
 
 const userHasData = () => {
 	const cookie =
+		getAdFreeCookie() ??
 		getCookie({ name: ACTION_REQUIRED_FOR_COOKIE }) ??
 		getCookie({ name: USER_FEATURES_EXPIRY_COOKIE }) ??
 		getCookie({ name: PAYING_MEMBER_COOKIE }) ??
 		getCookie({ name: RECURRING_CONTRIBUTOR_COOKIE }) ??
 		getCookie({ name: ONE_OFF_CONTRIBUTION_DATE_COOKIE }) ??
-		getCookie({ name: AD_FREE_USER_COOKIE }) ??
 		getCookie({ name: DIGITAL_SUBSCRIBER_COOKIE }) ??
 		getCookie({ name: HIDE_SUPPORT_MESSAGING_COOKIE });
 	return !!cookie;
@@ -64,70 +66,6 @@ const userHasData = () => {
 
 const accountDataUpdateWarning = (): string | null =>
 	getCookie({ name: ACTION_REQUIRED_FOR_COOKIE });
-
-const adFreeDataIsPresent = (): boolean => {
-	const cookieVal = getCookie({ name: AD_FREE_USER_COOKIE });
-	if (!cookieVal) return false;
-	return !Number.isNaN(parseInt(cookieVal, 10));
-};
-
-const setAdFreeCookieReason = (
-	reason: AdFreeCookieReasons,
-	expiry?: string,
-): void => {
-	const adFreeReasonString = localStorage.getItem(AD_FREE_COOKIE_REASON_LS);
-	const adFreeReason = JSON.parse(adFreeReasonString ?? '{}') as Partial<
-		Record<AdFreeCookieReasons, string>
-	>;
-
-	adFreeReason[reason] = expiry ?? timeInDaysFromNow(1);
-
-	localStorage.setItem(
-		AD_FREE_COOKIE_REASON_LS,
-		JSON.stringify(adFreeReason),
-	);
-};
-
-const unsetAdFreeCookieReason = (reason: AdFreeCookieReasons): void => {
-	const adFreeReasonString = localStorage.getItem(AD_FREE_COOKIE_REASON_LS);
-	const adFreeReason = JSON.parse(adFreeReasonString ?? '{}') as Partial<
-		Record<AdFreeCookieReasons, string>
-	>;
-
-	delete adFreeReason[reason];
-
-	localStorage.setItem(
-		AD_FREE_COOKIE_REASON_LS,
-		JSON.stringify(adFreeReason),
-	);
-};
-
-const isAdFreeCookieReasonFalseOrExpired = (): boolean => {
-	const adFreeReasonString = localStorage.getItem(AD_FREE_COOKIE_REASON_LS);
-	const adFreeReason = JSON.parse(adFreeReasonString ?? '{}') as Partial<
-		Record<AdFreeCookieReasons, string>
-	>;
-
-	const allExpired = Object.entries(AdFreeCookieReasons).every(
-		([, reason]) => {
-			const expiry = adFreeReason[reason];
-			if (expiry) {
-				const expiryTime = parseInt(expiry, 10);
-				const timeNow = new Date().getTime();
-				return timeNow >= expiryTime;
-			}
-			return true;
-		},
-	);
-
-	return allExpired;
-};
-
-const timeInDaysFromNow = (daysFromNow: number): string => {
-	const tmpDate = new Date();
-	tmpDate.setDate(tmpDate.getDate() + daysFromNow);
-	return tmpDate.getTime().toString();
-};
 
 /**
  * TODO: check that this validation is accurate
@@ -186,36 +124,21 @@ const persistResponse = (JsonResponse: UserFeaturesResponse) => {
 		});
 	}
 
-	if (
-		adFreeDataIsPresent() &&
-		!forcedAdFreeMode &&
-		!JsonResponse.contentAccess.digitalPack &&
-		isAdFreeCookieReasonFalseOrExpired()
-	) {
-		removeCookie({ name: AD_FREE_USER_COOKIE });
-		unsetAdFreeCookieReason(
-			AdFreeCookieReasons.AdFreeCookieReasonSubscriber,
-		);
-	}
-
 	if (JsonResponse.contentAccess.digitalPack) {
-		setCookie({ name: AD_FREE_USER_COOKIE, value: timeInDaysFromNow(2) });
-		setAdFreeCookieReason(
+		setAdFreeCookie(AdFreeCookieReasons.AdFreeCookieReasonSubscriber, 2);
+	} else if (adFreeDataIsPresent() && !forcedAdFreeMode) {
+		maybeUnsetAdFreeCookie(
 			AdFreeCookieReasons.AdFreeCookieReasonSubscriber,
-			timeInDaysFromNow(2),
 		);
 	}
 };
 
 const deleteOldData = (): void => {
 	// We expect adfree cookies to be cleaned up by the logout process, but what if the user's login simply times out?
-	unsetAdFreeCookieReason(AdFreeCookieReasons.AdFreeCookieReasonSubscriber);
+	maybeUnsetAdFreeCookie(AdFreeCookieReasons.AdFreeCookieReasonSubscriber);
 	removeCookie({ name: USER_FEATURES_EXPIRY_COOKIE });
 	removeCookie({ name: PAYING_MEMBER_COOKIE });
 	removeCookie({ name: RECURRING_CONTRIBUTOR_COOKIE });
-	if (isAdFreeCookieReasonFalseOrExpired()) {
-		removeCookie({ name: AD_FREE_USER_COOKIE });
-	}
 	removeCookie({ name: ACTION_REQUIRED_FOR_COOKIE });
 	removeCookie({ name: DIGITAL_SUBSCRIBER_COOKIE });
 	removeCookie({ name: HIDE_SUPPORT_MESSAGING_COOKIE });
@@ -241,24 +164,8 @@ const requestNewData = () =>
 		.then(persistResponse)
 		.catch(noop);
 
-const cookieIsExpiredOrMissing = (cookieName: string): boolean => {
-	const expiryDateFromCookie = getCookie({ name: cookieName });
-	if (!expiryDateFromCookie) return true;
-	const expiryTime = parseInt(expiryDateFromCookie, 10);
-	const timeNow = new Date().getTime();
-	return timeNow >= expiryTime;
-};
-
 const featuresDataIsOld = () =>
 	cookieIsExpiredOrMissing(USER_FEATURES_EXPIRY_COOKIE);
-
-const adFreeDataIsOld = (): boolean => {
-	const { switches } = window.guardian.config;
-	return (
-		Boolean(switches.adFreeStrictExpiryEnforcement) &&
-		cookieIsExpiredOrMissing(AD_FREE_USER_COOKIE)
-	);
-};
 
 const userNeedsNewFeatureData = (): boolean =>
 	featuresDataIsOld() || (adFreeDataIsPresent() && adFreeDataIsOld());
@@ -484,8 +391,5 @@ export {
 	ARTICLES_VIEWED_OPT_OUT_COOKIE,
 	CONTRIBUTIONS_REMINDER_SIGNED_UP,
 	canShowContributionsReminderFeature,
-	timeInDaysFromNow,
-	setAdFreeCookieReason,
-	unsetAdFreeCookieReason,
 	AdFreeCookieReasons,
 };
